Make auditing a bit more django-y.
We now create the audit log table as part of a migration.
This requires we have an oid and bigint field type. We possibly
could do these as part of a custom migration, but this way was
simpler.

We also only create the trigger function itself (in it's own schema).
This trigger function has been modified a fair bit, mostly to make
it simpler.

Turning on auditing for a given model is now done as part of a
django migration too: however there is a management command that
will write a migration file for you.

    $ ./manage.py audit_model <app-label> <model> [<model>...]

This file needs to be applied in order to turn on auditing.

There is currently a bug in django (https://code.djangoproject.com/ticket/22788)
which prevents this migration file from working as-is, it needs to be
edited after creation.