Audit to the same schema as the source.
This makes sense in a multi-tenancy setup, where the
audit table is duplicated per-schema, and the tenants
get a schema each.

Perhaps we could work around a single audit schema
by having the search path also contain the __audit
schema: in that case the audit table could be in there,
the public schema or the source table's schema.